# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "laravel-api脚手架",
        "description": "laravel-api脚手架",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://business-card.local",
            "description": "本地环境"
        },
        {
            "url": "http://dev.dev.com",
            "description": "开发环境"
        },
        {
            "url": "http://dev.stage.com",
            "description": "测试环境"
        },
        {
            "url": "",
            "description": "动态host"
        }
    ],
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "tags": [
                    "用户"
                ],
                "summary": "获取用户基本信息",
                "description": "获取用户基本信息",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::getUserBaseInfo",
                "responses": {
                    "400": {
                        "$ref": "#/components/responses/response_400"
                    },
                    "401": {
                        "$ref": "#/components/responses/response_401"
                    },
                    "404": {
                        "$ref": "#/components/responses/response_404"
                    },
                    "422": {
                        "$ref": "#/components/responses/response_422"
                    },
                    "429": {
                        "$ref": "#/components/responses/response_429"
                    },
                    "500": {
                        "$ref": "#/components/responses/response_500"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "UpdateItem",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response_schema"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/user_base_info_schema"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "用户"
                ],
                "summary": "编辑用户基本信息",
                "description": "编辑用户基本信息",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::postUserBaseInfo",
                "requestBody": {
                    "description": "swagger_user_request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user_base_info_schema"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "$ref": "#/components/responses/response_400"
                    },
                    "401": {
                        "$ref": "#/components/responses/response_401"
                    },
                    "404": {
                        "$ref": "#/components/responses/response_404"
                    },
                    "422": {
                        "$ref": "#/components/responses/response_422"
                    },
                    "429": {
                        "$ref": "#/components/responses/response_429"
                    },
                    "500": {
                        "$ref": "#/components/responses/response_500"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "UpdateItem",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response_schema"
                                        },
                                        {}
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "response_schema": {
                "properties": {
                    "status": {
                        "description": "状态: [T->成功, F->失败]",
                        "type": "string"
                    },
                    "message": {
                        "description": "接口消息",
                        "type": "string"
                    },
                    "code": {
                        "description": "错误码",
                        "type": "string"
                    },
                    "data": {
                        "description": "数据",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "user_base_info_schema": {
                "description": "用户基本信息",
                "properties": {
                    "mobile": {
                        "description": "手机号",
                        "type": "number"
                    },
                    "nick_name": {
                        "description": "昵称",
                        "type": "string"
                    },
                    "avatar_url": {
                        "description": "头像地址",
                        "type": "string"
                    },
                    "wechat_num": {
                        "description": "微信号",
                        "type": "string"
                    },
                    "duty": {
                        "description": "担任职务",
                        "type": "string"
                    },
                    "email": {
                        "description": "邮箱",
                        "type": "string"
                    },
                    "address": {
                        "description": "地址",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "用户",
                        "type": "number"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "response_400": {
                "description": "请求错误",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            },
            "response_401": {
                "description": "用户未授权",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            },
            "response_404": {
                "description": "页面不存在",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            },
            "response_422": {
                "description": "Unprocesable entity - [POST/PUT/PATCH] 当创建一个对象时，发生一个验证错误。",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            },
            "response_429": {
                "description": "Too Many Requests - [*] : 用户在给定的时间内发送了太多的请求",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            },
            "response_500": {
                "description": "INTERNAL SERVER ERROR - [*]：服务器发生错误，用户将无法判断发出的请求是否成功",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    },
...
}